Refactor: 엔드포인트 변경
