+++ .gitignore
@@ -1,4 +1,5 @@
-.env
+*.pem
+infra/.env
 HELP.md
 .gradle
 build/
